requirements.txt
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
piccolo[postgres]==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0

# Frontend
reflex==0.4.0

# Utilities
httpx==0.25.1
python-multipart==0.0.6
